BAN-71-UC-3 refactor: move comment component
